Minor changes to upload STIX to local instance
This was the original file uploaded for the local instance (not cloud).
Changed the get pass function to store the password as an environment variable. (for automated tasks such as cron jobs)
Fixed the description function to include sha-1 and md5 hashes. (these entries would break the original script)
Added an option to not verify certs (insecure) but can be allowed with other security mitigations in place.