stix_to_ecs v0.3.0
Rewritten code that supports both cloud and local instances. Included a cloud.json template and a local.json template to store credentials. added a few command switches in the help menu in line with the original content. Added an option to use an insecure flag for instances with self signed certs. (Ensure proper mitigation is in place before using the insecure flag).